+++ LazyEngine/MeshDataCPU/VertexBufferCPU/VertexBufferCPU.h
@@ -16,7 +16,7 @@ namespace LE {
 		
 		void addData(Primitives::Float32 vert);
 		//Function to get back a vector3 of x,y,z verts given an index into the m_data (Index should be mod 3) returns 0 otherwise
-		Primitives::Int16 getVertAtIndex(Primitives::Int32 index);
+		Primitives::Float32 getVertAtIndex(Primitives::Int32 index);
 		//Function to get number of vertices within this mesh
 		Primitives::Int32  getNumVerts();
 		//Function to get the mesh name
